sandbox: add env:flex configuration file
vm: true is going away and less stable.
Add an alternate configuration file that uses the new App Engine
flexible environment.

NOTE: the new environment currently has vastly different network
configuration, so it's important it's hosted at a different URL for now.
The module name is changed to "sandbox-flex" to reflect this.

Change-Id: Iec3ed9bdb1d61f6aafd26350a01ad268adf3194e
Reviewed-on: https://go-review.googlesource.com/34559
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>