playground: More verbose error message
The existing error message when accessing a code snippet from an
unknown country (App Engine detected as "ZZ") have been subject for
confusion several times. I've seen many users thinking that the
particular code snippet they are trying to access contains code that
isn't legal to share, when it is only App Engine that fails to identify
their origin. Most often this is because they are behind a proxy.

This change gives a more helpful message. It is also formated in HTML as
the previous used html.Error() that showed that message as text,
verbatim.

Change-Id: Idb8598f933bb46b79ea417310e965f3fa774f0ce
Reviewed-on: https://go-review.googlesource.com/48930
Reviewed-by: Andrew Bonventre <andybons@golang.org>