frontend: increase hash length to 11 bytes
In changing the hash function and salt in https://golang.org/cl/84977
there is a (very small) chance that collisions may occur with newly
created snippets. If Bob created a snippet using the previous hash,
but Alice created a new snippet using the new method that happened to
have the same 10-byte string ID, then it would overwrite Bob’s snippet.

Increase the length by 1 to avoid this and to also clearly delineate
which are saved using the old method and which are using the new one.

Change-Id: Ide36b76becb3b105899d71d7965291ec51961278
Reviewed-on: https://go-review.googlesource.com/85035
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>