playground: support examples as tests
Playground handles examples in the same way as "go test" does.

If the program contains examples and does not import the "testing"
package, the playground injects the import without changing line numbers.
This is to prevent a mismatch of highlighted error lines on the
web page. The resulting import code is not properly formatted,
but it does not have to be since it is not meant to be read by humans.

Fixes golang/go#24311

Change-Id: I681a22d40d191fe706dfa9949421f4d3829031a4
Reviewed-on: https://go-review.googlesource.com/99595
Reviewed-by: Yury Smolsky <yury@smolsky.by>
Reviewed-by: Josh Bleecher Snyder <josharian@gmail.com>