playground: interleave standard output and error correctly
The change reconstructs the correct interleaving of the standard
output and error. Playground buffers those streams separetely
and performs a sorted merge.

The change depends on the change https://golang.org/cl/105235,
which makes Playground timestamps change when the stream fd changes.
It's recommended to merge this change after the CL 105235 went into
go release, but it can be picked with CL 106216 until that happens.

Fixes golang/go#24615
Fixes golang/go#24659

Change-Id: I76f8815c47e8961d5c435e55c761c7a7bf3920ce
Reviewed-on: https://go-review.googlesource.com/105496
Reviewed-by: Andrew Bonventre <andybons@golang.org>
Reviewed-by: Bryan C. Mills <bcmills@google.com>