README: document the need to increase cloud build timeout
It's unfortunate to introduce an extra step of setting a global
configuration value to the deploy process. It would be better if
this could be an application-scoped setting specified in the app.yaml
configuration. However, that doesn't seem to be possible, and after
reading through documentation and waiting to hear any suggestions at
golang/go#28046, I have not become aware of a better solution.

Document this approach for now, and if we learn of a better solution
in the future, this can be improved then.

Fixes golang/go#28046

Change-Id: I2e9b7bbd38aa8313ff6eab0bf99dd2fe48f57544
Reviewed-on: https://go-review.googlesource.com/c/playground/+/164577
Reviewed-by: Andrew Bonventre <andybons@golang.org>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>