playground: build playground Go from source
This changes the playground Dockerfile to build playground from source.
Cloud Builds for playground are triggered when a new Go release is
tagged. Go release tags are published to repositories before binary
releases are published. This change fetches the tagged source in order
to build Go, as a tagged binary release may not be available yet.

A binary Go release is still downloaded to be used as the
GOROOT_BOOTSTRAP.

The build steps use maintq to fetch the latest version of Go. maintq
currently only outputs in textproto format, requiring a step to parse
the git tag. It might be worth adding a flag or special command to
maintq to simplify the output to clarify this step of the release
process. It might also be worthwhile to move this step into the
Makefile, or a checked-in build script.

The trigger steps were tested on Cloud Build without the deployment
step.

Updates golang/go#32606

Change-Id: I811be6aa6b5fb53fce4491e4b4ed235a0f3e62cb
Reviewed-on: https://go-review.googlesource.com/c/playground/+/183403
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>