playground: push image in trigger before deploying
Removes post-build trigger step of pushing image, and adds a push step
step before AppEngine deployment. This ensures that the image we just
built is pushed before asking AppEngine to deploy it.

This is documented at:
https://cloud.google.com/cloud-build/docs/configuring-builds/build-test-deploy-artifacts

Cloud Build is authorized to push images to gcr.io. This change was
tested by submitting a build with this configuration without the
deployment step.

Updates golang/go#32606

Change-Id: I4e7cc242566378e4ca6d52244fb81bf410e79314
Reviewed-on: https://go-review.googlesource.com/c/playground/+/185339
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>