playground: slurp binary to memory in gvisor mode
Slurp the binary to memory before POSTing it. This simplifies retries
a bit (doesn't need to re-open the file) and sets a Content-Length
header, and checks the file size early, rather than having the backend
complain.

None of this should matter for correctness (and unnecessarily uses
more memory than before), but we're sometimes seeing timeout errors
followed by this frontend returning (and caching!) an empty response
where the backend never sees a binary arrive, so this is somewhat a
shot in the dark for now. Something's up with our internal load
balancer, perhaps. Still debugging.

Also update/add some comments.

Updates golang/go#25224

Change-Id: I12f08db8ac77743b555fe7ef56bccf8cbc45742c
Reviewed-on: https://go-review.googlesource.com/c/playground/+/215058
Reviewed-by: Alexander Rakoczy <alex@golang.org>