sandbox: instrument HTTP handlers with StackDriver
This change adds OpenCensus's HTTP instrumentation to the sandbox. In
development mode, it exposes a prometheus metrics interface on /statusz.
This is the first in a series of CLs to add instrumentation to different
parts of the sandbox to help investigate instability issues. For now,
reporting metrics around our responses per handler to StackDriver will
be helpful.

OpenTelemetry would be preferable, as it is the successor of OpenCensus,
however the StackDriver integration is not quite done.

Updates golang/go#25224
Updates golang/go#38530

Change-Id: I600fd695bb66c8bee16bc0b778d51930f4cdd476
Reviewed-on: https://go-review.googlesource.com/c/playground/+/229679
Run-TryBot: Alexander Rakoczy <alex@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Carlos Amedee <carlos@golang.org>